+++ Editor/Settings/EditorAnnotationSettingsProvider.cs
@@ -1,5 +1,5 @@
 using Lin.Editor.Annotation.Asset;
-using Lin.Editor.Annotation.Toolbar.Element;
+using Lin.Editor.Toolbar.Element;
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;