+++ Editor/Toolbar/Element/EVisibleMode.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Lin.Editor.Annotation.Toolbar.Element
+namespace Lin.Editor.Toolbar.Element
 {
     [Flags]
     public enum EVisibleMode