+++ Editor/Toolbar/Element/IToolbarElement.cs
@@ -1,6 +1,6 @@
 using UnityEngine.UIElements;
 
-namespace Lin.Editor.Annotation.Toolbar.Element
+namespace Lin.Editor.Toolbar.Element
 {
     public interface IToolbarElement
     {